+++ docs/changes.rst
@@ -12,6 +12,7 @@ Not yet released.
 * Use content compression for file downloads.
 * Improved user experience on importing from ZIP file.
 * Improved detection of file format for uploads.
+* Avoid duplicate pull requests on Pagure.
 
 Weblate 4.4
 -----------